requirements.txt
pyttsx3>=2.90
SpeechRecognition>=3.10
python-dotenv>=0.19.0
flask>=3.0
flask-cors>=4.0
psutil>=5.9
requests>=2.28
opencv-python>=4.8
waitress>=3.0
Pillow>=10.0
google-generativeai>=0.4.0
